Add market dropdown as seperate component - And have it render on it's own row
