+++ translations/en.json
@@ -554,6 +554,7 @@
 				"rate": "24H funding rate"
 			},
 			"trade": {
+				"market": "market",
 				"title": "trade",
 				"input": {
 					"buy": "buy",
